assimp-rs: Add read_memory_with_hint
Add new reading function which allows specifying the "hint" to the
read_file call. This allows the file extensions to be passed in, which
greatly affects the probing algorithm in assimp.

Without the hint present, reading a .obj file from memory was
accidentally probing positive with the Blender importer, due to the word
'blender' in the header comments of the .obj.  This is because reading
from memory lacks a filename altogether, which let that probe hit early.

Fix by exposing this new loading call that extends the interface.